requirements.txt
# torch==1.12.1
transformers==4.30.2
lightning==2.0.5
Pillow==9.4.0
numpy==1.23.5
av==10.0.0
pandas==1.5.3
deepspeed==0.10.0
tensorboard==2.12.0
protobuf==3.20
sentencepiece
opencv-python
gradio
einops
imageio
torchvision